classe code cleanup

+++ app/Http/Controllers/ClasseController.php
@@ -2,19 +2,19 @@
 
 namespace App\Http\Controllers;
 
+use App\Exceptions\NotFoundException;
+use App\Exceptions\IsExisteExcetion;
 use App\Http\Requests\StoreClasseRequest;
+use App\Http\Requests\UpdateClasseRequest;
+use App\Http\Resources\ClasseResource\ClasseResource;
 use App\Http\Resources\ClasseResource\ClasseResourceCollection;
 use App\Models\Classe;
 use App\Services\ClasseService;
-use Illuminate\Http\Request;
 
 class ClasseController extends Controller
 {
-    private $ClasseService;
-
-    public function __construct(ClasseService $ClasseService)
+    public function __construct(private ClasseService $classeService)
     {
-        $this->ClasseService = $ClasseService;
     }
 
 
@@ -43,7 +43,17 @@ class ClasseController extends Controller
      */
     public function store(StoreClasseRequest $request)
     {
-        $this->ClasseService->store($request->validated());
+        $niveau = $this->classeService->checkniveauexiste($request['niveau_id']);
+        if (is_null($niveau)) {
+            throw new NotFoundException(['code' => -3, 'message' => ' niveau not found']);
+        }
+
+        $classe = $this->classeService->checkClasseNameExiste($request['name']);
+        if (isset($classe)) {
+            throw new IsExisteExcetion(['code' => -1, 'name' => ' classe']);
+        }
+
+        $this->classeService->store($request->validated());
         return response(['success' => 1, 'message' => 'classe is create'], 201);
     }
 
@@ -56,11 +66,14 @@ class ClasseController extends Controller
      */
     public function show($id)
     {
-        $classe = $this->ClasseService->checkClasseNotExiste($id);
+        $classe = $this->classeService->checkClasseNotExiste($id);
+        if (is_null($classe)) {
+            throw new NotFoundException(['code' => -2, 'message' => ' classe not found']);
+        }
         return response(
             [
                 'success' => 1,
-                'data' => new ClasseResourceCollection($classe)
+                'data' => new ClasseResource($classe)
             ],
             201
         );
@@ -74,9 +87,22 @@ class ClasseController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(UpdateClasseRequest $request, $id)
     {
-        $this->ClasseService->update($request->validated(), $id);
+        $classe = $this->classeService->checkClasseNotExiste($id);
+        if (is_null($classe)) {
+            throw new NotFoundException(['code' => -3, 'message' => 'classe not found']);
+        }
+        $niveau = $this->classeService->checkniveauexiste($request['niveau_id']);
+        if (is_null($niveau)) {
+            throw new NotFoundException(['code' => -3, 'message' => 'niveau not found']);
+        }
+        $name = $this->classeService->checkClasseNameIDExiste($request['name'], $id);
+        if (isset($name)) {
+            throw new IsExisteExcetion(['code' => -1, 'name' => 'niveau name']);
+        }
+
+        $this->classeService->update($request->validated(), $id, $classe);
         return response(['success' => 1, 'message' => 'classe is updated'], 201);
     }
 
@@ -89,7 +115,10 @@ class ClasseController extends Controller
      */
     public function destroy($id)
     {
-        $classe = $this->ClasseService->checkClasseNotExiste($id);
+        $classe = $this->classeService->checkClasseNotExiste($id);
+        if (is_null($classe)) {
+            throw new NotFoundException(['code' => -3, 'message' => 'classe not found']);
+        }
         $classe->delete();
         return response(['success' => 1, 'message' => 'classe is deleted'], 201);
     }
